Split markdown remote image loader
